requirements.txt
asgiref
dj-database-url
Django
django-cors-headers
djangorestframework
gunicorn
importlib-metadata
packaging
psycopg2-binary
python-dotenv
pytz
sqlparse
typing-extensions
zipp